backend/requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
pydantic>=2.5.0
websockets>=12.0
python-multipart>=0.0.6

# Computer Vision dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0